\name{dataS}
\alias{dataS}
\docType{data}
\title{dataS
}
\description{
Dataset S for usage examples.
}
\usage{data("dataS")}
\format{
  A data frame with 1000 observations on the following 305 variables.
  \describe{
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{x3}}{a numeric vector}
    \item{\code{x4}}{a numeric vector}
    \item{\code{x5}}{a numeric vector}
    \item{\code{x6}}{a numeric vector}
    \item{\code{x7}}{a numeric vector}
    \item{\code{x8}}{a numeric vector}
    \item{\code{x9}}{a numeric vector}
    \item{\code{x10}}{a numeric vector}
    \item{\code{x11}}{a numeric vector}
    \item{\code{x12}}{a numeric vector}
    \item{\code{x13}}{a numeric vector}
    \item{\code{x14}}{a numeric vector}
    \item{\code{x15}}{a numeric vector}
    \item{\code{x16}}{a numeric vector}
    \item{\code{x17}}{a numeric vector}
    \item{\code{x18}}{a numeric vector}
    \item{\code{x19}}{a numeric vector}
    \item{\code{x20}}{a numeric vector}
    \item{\code{x21}}{a numeric vector}
    \item{\code{x22}}{a numeric vector}
    \item{\code{x23}}{a numeric vector}
    \item{\code{x24}}{a numeric vector}
    \item{\code{x25}}{a numeric vector}
    \item{\code{x26}}{a numeric vector}
    \item{\code{x27}}{a numeric vector}
    \item{\code{x28}}{a numeric vector}
    \item{\code{x29}}{a numeric vector}
    \item{\code{x30}}{a numeric vector}
    \item{\code{x31}}{a numeric vector}
    \item{\code{x32}}{a numeric vector}
    \item{\code{x33}}{a numeric vector}
    \item{\code{x34}}{a numeric vector}
    \item{\code{x35}}{a numeric vector}
    \item{\code{x36}}{a numeric vector}
    \item{\code{x37}}{a numeric vector}
    \item{\code{x38}}{a numeric vector}
    \item{\code{x39}}{a numeric vector}
    \item{\code{x40}}{a numeric vector}
    \item{\code{x41}}{a numeric vector}
    \item{\code{x42}}{a numeric vector}
    \item{\code{x43}}{a numeric vector}
    \item{\code{x44}}{a numeric vector}
    \item{\code{x45}}{a numeric vector}
    \item{\code{x46}}{a numeric vector}
    \item{\code{x47}}{a numeric vector}
    \item{\code{x48}}{a numeric vector}
    \item{\code{x49}}{a numeric vector}
    \item{\code{x50}}{a numeric vector}
    \item{\code{x51}}{a numeric vector}
    \item{\code{x52}}{a numeric vector}
    \item{\code{x53}}{a numeric vector}
    \item{\code{x54}}{a numeric vector}
    \item{\code{x55}}{a numeric vector}
    \item{\code{x56}}{a numeric vector}
    \item{\code{x57}}{a numeric vector}
    \item{\code{x58}}{a numeric vector}
    \item{\code{x59}}{a numeric vector}
    \item{\code{x60}}{a numeric vector}
    \item{\code{x61}}{a numeric vector}
    \item{\code{x62}}{a numeric vector}
    \item{\code{x63}}{a numeric vector}
    \item{\code{x64}}{a numeric vector}
    \item{\code{x65}}{a numeric vector}
    \item{\code{x66}}{a numeric vector}
    \item{\code{x67}}{a numeric vector}
    \item{\code{x68}}{a numeric vector}
    \item{\code{x69}}{a numeric vector}
    \item{\code{x70}}{a numeric vector}
    \item{\code{x71}}{a numeric vector}
    \item{\code{x72}}{a numeric vector}
    \item{\code{x73}}{a numeric vector}
    \item{\code{x74}}{a numeric vector}
    \item{\code{x75}}{a numeric vector}
    \item{\code{x76}}{a numeric vector}
    \item{\code{x77}}{a numeric vector}
    \item{\code{x78}}{a numeric vector}
    \item{\code{x79}}{a numeric vector}
    \item{\code{x80}}{a numeric vector}
    \item{\code{x81}}{a numeric vector}
    \item{\code{x82}}{a numeric vector}
    \item{\code{x83}}{a numeric vector}
    \item{\code{x84}}{a numeric vector}
    \item{\code{x85}}{a numeric vector}
    \item{\code{x86}}{a numeric vector}
    \item{\code{x87}}{a numeric vector}
    \item{\code{x88}}{a numeric vector}
    \item{\code{x89}}{a numeric vector}
    \item{\code{x90}}{a numeric vector}
    \item{\code{x91}}{a numeric vector}
    \item{\code{x92}}{a numeric vector}
    \item{\code{x93}}{a numeric vector}
    \item{\code{x94}}{a numeric vector}
    \item{\code{x95}}{a numeric vector}
    \item{\code{x96}}{a numeric vector}
    \item{\code{x97}}{a numeric vector}
    \item{\code{x98}}{a numeric vector}
    \item{\code{x99}}{a numeric vector}
    \item{\code{x100}}{a numeric vector}
    \item{\code{m1}}{a numeric vector}
    \item{\code{m2}}{a numeric vector}
    \item{\code{m3}}{a numeric vector}
    \item{\code{m4}}{a numeric vector}
    \item{\code{m5}}{a numeric vector}
    \item{\code{m6}}{a numeric vector}
    \item{\code{m7}}{a numeric vector}
    \item{\code{m8}}{a numeric vector}
    \item{\code{m9}}{a numeric vector}
    \item{\code{m10}}{a numeric vector}
    \item{\code{m11}}{a numeric vector}
    \item{\code{m12}}{a numeric vector}
    \item{\code{m13}}{a numeric vector}
    \item{\code{m14}}{a numeric vector}
    \item{\code{m15}}{a numeric vector}
    \item{\code{m16}}{a numeric vector}
    \item{\code{m17}}{a numeric vector}
    \item{\code{m18}}{a numeric vector}
    \item{\code{m19}}{a numeric vector}
    \item{\code{m20}}{a numeric vector}
    \item{\code{m21}}{a numeric vector}
    \item{\code{m22}}{a numeric vector}
    \item{\code{m23}}{a numeric vector}
    \item{\code{m24}}{a numeric vector}
    \item{\code{m25}}{a numeric vector}
    \item{\code{m26}}{a numeric vector}
    \item{\code{m27}}{a numeric vector}
    \item{\code{m28}}{a numeric vector}
    \item{\code{m29}}{a numeric vector}
    \item{\code{m30}}{a numeric vector}
    \item{\code{m31}}{a numeric vector}
    \item{\code{m32}}{a numeric vector}
    \item{\code{m33}}{a numeric vector}
    \item{\code{m34}}{a numeric vector}
    \item{\code{m35}}{a numeric vector}
    \item{\code{m36}}{a numeric vector}
    \item{\code{m37}}{a numeric vector}
    \item{\code{m38}}{a numeric vector}
    \item{\code{m39}}{a numeric vector}
    \item{\code{m40}}{a numeric vector}
    \item{\code{m41}}{a numeric vector}
    \item{\code{m42}}{a numeric vector}
    \item{\code{m43}}{a numeric vector}
    \item{\code{m44}}{a numeric vector}
    \item{\code{m45}}{a numeric vector}
    \item{\code{m46}}{a numeric vector}
    \item{\code{m47}}{a numeric vector}
    \item{\code{m48}}{a numeric vector}
    \item{\code{m49}}{a numeric vector}
    \item{\code{m50}}{a numeric vector}
    \item{\code{m51}}{a numeric vector}
    \item{\code{m52}}{a numeric vector}
    \item{\code{m53}}{a numeric vector}
    \item{\code{m54}}{a numeric vector}
    \item{\code{m55}}{a numeric vector}
    \item{\code{m56}}{a numeric vector}
    \item{\code{m57}}{a numeric vector}
    \item{\code{m58}}{a numeric vector}
    \item{\code{m59}}{a numeric vector}
    \item{\code{m60}}{a numeric vector}
    \item{\code{m61}}{a numeric vector}
    \item{\code{m62}}{a numeric vector}
    \item{\code{m63}}{a numeric vector}
    \item{\code{m64}}{a numeric vector}
    \item{\code{m65}}{a numeric vector}
    \item{\code{m66}}{a numeric vector}
    \item{\code{m67}}{a numeric vector}
    \item{\code{m68}}{a numeric vector}
    \item{\code{m69}}{a numeric vector}
    \item{\code{m70}}{a numeric vector}
    \item{\code{m71}}{a numeric vector}
    \item{\code{m72}}{a numeric vector}
    \item{\code{m73}}{a numeric vector}
    \item{\code{m74}}{a numeric vector}
    \item{\code{m75}}{a numeric vector}
    \item{\code{m76}}{a numeric vector}
    \item{\code{m77}}{a numeric vector}
    \item{\code{m78}}{a numeric vector}
    \item{\code{m79}}{a numeric vector}
    \item{\code{m80}}{a numeric vector}
    \item{\code{m81}}{a numeric vector}
    \item{\code{m82}}{a numeric vector}
    \item{\code{m83}}{a numeric vector}
    \item{\code{m84}}{a numeric vector}
    \item{\code{m85}}{a numeric vector}
    \item{\code{m86}}{a numeric vector}
    \item{\code{m87}}{a numeric vector}
    \item{\code{m88}}{a numeric vector}
    \item{\code{m89}}{a numeric vector}
    \item{\code{m90}}{a numeric vector}
    \item{\code{m91}}{a numeric vector}
    \item{\code{m92}}{a numeric vector}
    \item{\code{m93}}{a numeric vector}
    \item{\code{m94}}{a numeric vector}
    \item{\code{m95}}{a numeric vector}
    \item{\code{m96}}{a numeric vector}
    \item{\code{m97}}{a numeric vector}
    \item{\code{m98}}{a numeric vector}
    \item{\code{m99}}{a numeric vector}
    \item{\code{m100}}{a numeric vector}
    \item{\code{f1}}{a numeric vector}
    \item{\code{f2}}{a numeric vector}
    \item{\code{f3}}{a numeric vector}
    \item{\code{f4}}{a numeric vector}
    \item{\code{f5}}{a numeric vector}
    \item{\code{f6}}{a numeric vector}
    \item{\code{f7}}{a numeric vector}
    \item{\code{f8}}{a numeric vector}
    \item{\code{f9}}{a numeric vector}
    \item{\code{f10}}{a numeric vector}
    \item{\code{f11}}{a numeric vector}
    \item{\code{f12}}{a numeric vector}
    \item{\code{f13}}{a numeric vector}
    \item{\code{f14}}{a numeric vector}
    \item{\code{f15}}{a numeric vector}
    \item{\code{f16}}{a numeric vector}
    \item{\code{f17}}{a numeric vector}
    \item{\code{f18}}{a numeric vector}
    \item{\code{f19}}{a numeric vector}
    \item{\code{f20}}{a numeric vector}
    \item{\code{f21}}{a numeric vector}
    \item{\code{f22}}{a numeric vector}
    \item{\code{f23}}{a numeric vector}
    \item{\code{f24}}{a numeric vector}
    \item{\code{f25}}{a numeric vector}
    \item{\code{f26}}{a numeric vector}
    \item{\code{f27}}{a numeric vector}
    \item{\code{f28}}{a numeric vector}
    \item{\code{f29}}{a numeric vector}
    \item{\code{f30}}{a numeric vector}
    \item{\code{f31}}{a numeric vector}
    \item{\code{f32}}{a numeric vector}
    \item{\code{f33}}{a numeric vector}
    \item{\code{f34}}{a numeric vector}
    \item{\code{f35}}{a numeric vector}
    \item{\code{f36}}{a numeric vector}
    \item{\code{f37}}{a numeric vector}
    \item{\code{f38}}{a numeric vector}
    \item{\code{f39}}{a numeric vector}
    \item{\code{f40}}{a numeric vector}
    \item{\code{f41}}{a numeric vector}
    \item{\code{f42}}{a numeric vector}
    \item{\code{f43}}{a numeric vector}
    \item{\code{f44}}{a numeric vector}
    \item{\code{f45}}{a numeric vector}
    \item{\code{f46}}{a numeric vector}
    \item{\code{f47}}{a numeric vector}
    \item{\code{f48}}{a numeric vector}
    \item{\code{f49}}{a numeric vector}
    \item{\code{f50}}{a numeric vector}
    \item{\code{f51}}{a numeric vector}
    \item{\code{f52}}{a numeric vector}
    \item{\code{f53}}{a numeric vector}
    \item{\code{f54}}{a numeric vector}
    \item{\code{f55}}{a numeric vector}
    \item{\code{f56}}{a numeric vector}
    \item{\code{f57}}{a numeric vector}
    \item{\code{f58}}{a numeric vector}
    \item{\code{f59}}{a numeric vector}
    \item{\code{f60}}{a numeric vector}
    \item{\code{f61}}{a numeric vector}
    \item{\code{f62}}{a numeric vector}
    \item{\code{f63}}{a numeric vector}
    \item{\code{f64}}{a numeric vector}
    \item{\code{f65}}{a numeric vector}
    \item{\code{f66}}{a numeric vector}
    \item{\code{f67}}{a numeric vector}
    \item{\code{f68}}{a numeric vector}
    \item{\code{f69}}{a numeric vector}
    \item{\code{f70}}{a numeric vector}
    \item{\code{f71}}{a numeric vector}
    \item{\code{f72}}{a numeric vector}
    \item{\code{f73}}{a numeric vector}
    \item{\code{f74}}{a numeric vector}
    \item{\code{f75}}{a numeric vector}
    \item{\code{f76}}{a numeric vector}
    \item{\code{f77}}{a numeric vector}
    \item{\code{f78}}{a numeric vector}
    \item{\code{f79}}{a numeric vector}
    \item{\code{f80}}{a numeric vector}
    \item{\code{f81}}{a numeric vector}
    \item{\code{f82}}{a numeric vector}
    \item{\code{f83}}{a numeric vector}
    \item{\code{f84}}{a numeric vector}
    \item{\code{f85}}{a numeric vector}
    \item{\code{f86}}{a numeric vector}
    \item{\code{f87}}{a numeric vector}
    \item{\code{f88}}{a numeric vector}
    \item{\code{f89}}{a numeric vector}
    \item{\code{f90}}{a numeric vector}
    \item{\code{f91}}{a numeric vector}
    \item{\code{f92}}{a numeric vector}
    \item{\code{f93}}{a numeric vector}
    \item{\code{f94}}{a numeric vector}
    \item{\code{f95}}{a numeric vector}
    \item{\code{f96}}{a numeric vector}
    \item{\code{f97}}{a numeric vector}
    \item{\code{f98}}{a numeric vector}
    \item{\code{f99}}{a numeric vector}
    \item{\code{f100}}{a numeric vector}
    \item{\code{weight}}{a numeric vector}
    \item{\code{height}}{a numeric vector}
    \item{\code{fev}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{gender}}{a numeric vector}
  }
}

\keyword{dataset}
